[chore] Merge batch_sander into the queue_sander
