+++ exporter/exporterhelper/internal/queue_sender.go
@@ -14,6 +14,14 @@ import (
 	"go.opentelemetry.io/collector/exporter/exporterqueue"
 	"go.opentelemetry.io/collector/exporter/internal"
 	"go.opentelemetry.io/collector/exporter/internal/queue"
+	"go.opentelemetry.io/collector/featuregate"
+)
+
+var usePullingBasedExporterQueueBatcher = featuregate.GlobalRegistry().MustRegister(
+	"exporter.UsePullingBasedExporterQueueBatcher",
+	featuregate.StageAlpha,
+	featuregate.WithRegisterFromVersion("v0.115.0"),
+	featuregate.WithRegisterDescription("if set to true, turns on the pulling-based exporter queue bathcer"),
 )
 
 // QueueConfig defines configuration for queueing batches before sending to the consumerSender.
@@ -67,6 +75,7 @@ func (qCfg *QueueConfig) Validate() error {
 type QueueSender struct {
 	queue   exporterqueue.Queue[internal.Request]
 	batcher component.Component
+	bs      *BatchSender
 }
 
 func NewQueueSender(
@@ -77,6 +86,39 @@ func NewQueueSender(
 	exportFailureMessage string,
 	next Sender[internal.Request],
 ) (*QueueSender, error) {
+	if !usePullingBasedExporterQueueBatcher.IsEnabled() {
+		concurrencyLimit := int64(0)
+		if qCfg.Enabled {
+			concurrencyLimit = int64(qCfg.NumConsumers)
+		}
+
+		var bs *BatchSender
+		if bCfg.Enabled {
+			bs = NewBatchSender(bCfg, qSet.ExporterSettings, concurrencyLimit, next)
+			next = bs
+		}
+
+		exportFunc := func(ctx context.Context, req internal.Request) error {
+			// Have to read the number of items before sending the request since the request can
+			// be modified by the downstream components like the batcher.
+			itemsCount := req.ItemsCount()
+			err := next.Send(ctx, req)
+			if err != nil {
+				qSet.ExporterSettings.Logger.Error("Exporting failed. Dropping data."+exportFailureMessage,
+					zap.Error(err), zap.Int("dropped_items", itemsCount))
+			}
+			return err
+		}
+
+		q, err := newObsQueue(qSet, qf(context.Background(), qSet, qCfg, func(ctx context.Context, req internal.Request, done exporterqueue.Done) {
+			done.OnDone(exportFunc(ctx, req))
+		}))
+		if err != nil {
+			return nil, err
+		}
+		return &QueueSender{queue: q, bs: bs}, nil
+	}
+
 	exportFunc := func(ctx context.Context, req internal.Request) error {
 		// Have to read the number of items before sending the request since the request can
 		// be modified by the downstream components like the batcher.
@@ -87,15 +129,6 @@ func NewQueueSender(
 				zap.Error(err), zap.Int("dropped_items", itemsCount))
 		}
 		return err
-	}
-	if !usePullingBasedExporterQueueBatcher.IsEnabled() {
-		q, err := newObsQueue(qSet, qf(context.Background(), qSet, qCfg, func(ctx context.Context, req internal.Request, done exporterqueue.Done) {
-			done.OnDone(exportFunc(ctx, req))
-		}))
-		if err != nil {
-			return nil, err
-		}
-		return &QueueSender{queue: q}, nil
 	}
 
 	b, err := queue.NewBatcher(bCfg, exportFunc, qCfg.NumConsumers)
@@ -116,6 +149,13 @@ func NewQueueSender(
 
 // Start is invoked during service startup.
 func (qs *QueueSender) Start(ctx context.Context, host component.Host) error {
+	if qs.bs != nil {
+		// If no error then start the BatchSender.
+		if err := qs.bs.Start(ctx, host); err != nil {
+			return err
+		}
+	}
+
 	if err := qs.queue.Start(ctx, host); err != nil {
 		return err
 	}
@@ -129,9 +169,15 @@ func (qs *QueueSender) Start(ctx context.Context, host component.Host) error {
 
 // Shutdown is invoked during service shutdown.
 func (qs *QueueSender) Shutdown(ctx context.Context) error {
+	var err error
+	// Then shutdown the batch sender
+	if qs.bs != nil {
+		err = errors.Join(err, qs.bs.Shutdown(ctx))
+	}
+
 	// Stop the queue and batcher, this will drain the queue and will call the retry (which is stopped) that will only
 	// try once every request.
-	err := qs.queue.Shutdown(ctx)
+	err = errors.Join(err, qs.queue.Shutdown(ctx))
 	if usePullingBasedExporterQueueBatcher.IsEnabled() {
 		return errors.Join(err, qs.batcher.Shutdown(ctx))
 	}
